Adding `pinned` badge to pinned comments
I added a badge to indicated when a comment is in the pinned section.
This is because a pinned comment apears twice, once in the Pinned
section and in the regular section. The badge allows to diferentiate its
ocurrences.

To do this, the parsing of the comments was split into each section
and a `pinned` attribute was added to the comment tuple.